+++ packages/patrol/lib/src/common.dart
@@ -1,6 +1,7 @@
 import 'dart:developer';
 import 'dart:io' as io;
 
+import 'package:boolean_selector/boolean_selector.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:meta/meta.dart';
@@ -170,12 +171,15 @@ DartGroupEntry createDartTestGroup(
   String name = '',
   int level = 0,
   int maxTestCaseLength = global_state.maxTestLength,
+  String? tags,
+  String? excludeTags,
 }) {
   final groupDTO = DartGroupEntry(
     name: name,
     type: GroupEntryType.group,
     entries: [],
     skip: parentGroup.metadata.skip,
+    tags: parentGroup.metadata.tags.toList(),
   );
 
   for (final entry in parentGroup.entries) {
@@ -204,14 +208,34 @@ DartGroupEntry createDartTestGroup(
           throw StateError('Test is not allowed to be defined at level $level');
         }
 
+        if (tags != null) {
+          final includeTagsSelector = BooleanSelector.parse(tags);
+
+          // If the user provided tags, skip tests that don't match all of them.
+          if (!includeTagsSelector.evaluate(entry.metadata.tags.contains)) {
+            continue;
+          }
+        }
+
+        if (excludeTags != null) {
+          final excludeTagsSelector = BooleanSelector.parse(excludeTags);
+
+          // Skip tests that do match any tags the user wants to exclude.
+          if (excludeTagsSelector.evaluate(entry.metadata.tags.contains)) {
+            continue;
+          }
+        }
+
         groupDTO.entries.add(
           DartGroupEntry(
             name: name,
             type: GroupEntryType.test,
             entries: [],
             skip: entry.metadata.skip,
+            tags: entry.metadata.tags.toList(),
           ),
         );
+
       case Group _:
         groupDTO.entries.add(
           createDartTestGroup(
@@ -219,6 +243,8 @@ DartGroupEntry createDartTestGroup(
             name: name,
             level: level + 1,
             maxTestCaseLength: maxTestCaseLength,
+            tags: tags,
+            excludeTags: excludeTags,
           ),
         );
     }