+++ packages/patrol_cli/lib/src/commands/test.dart
@@ -55,6 +55,8 @@ class TestCommand extends PatrolCommand {
     usesWaitOption();
     usesPortOptions();
     useCoverageOptions();
+    usesTagsOption();
+    usesExcludeTagsOption();
 
     usesUninstallOption();
 
@@ -107,9 +109,17 @@ class TestCommand extends PatrolCommand {
       _logger.detail('Received test target: $t');
     }
 
+    final tags = stringArg('tags');
+    final excludeTags = stringArg('exclude-tags');
+    if (tags != null) {
+      _logger.detail('Received tag(s): $tags');
+    }
+    if (excludeTags != null) {
+      _logger.detail('Received exclude tag(s): $excludeTags');
+    }
     final entrypoint = _testBundler.bundledTestFile;
     if (boolArg('generate-bundle')) {
-      _testBundler.createTestBundle(targets);
+      _testBundler.createTestBundle(targets, tags, excludeTags);
     }
 
     final androidFlavor = stringArg('flavor') ?? config.android.flavor;