+++ packages/patrol_cli/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.1.0
+- Add `tags` and `exclude-tags`. (#2286)
+- Run `flutter build apk --config-only` during android build.(#2293)
+
+This version requires version 3.10.0 of `patrol` package.
+
 ## 3.0.1
 
 - Fallback to read `java` version from `JAVA_HOME` when `flutter doctor` doesn't print any.